requirements.txt
asgiref==3.8.1
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.7.0
django-cors-headers==4.7.0
django-formtools==2.5.1
django-otp==1.6.0
django-phonenumber-field==8.1.0
django-two-factor-auth==1.17.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
ecdsa==0.19.1
pillow==11.1.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
PyJWT==2.9.0
pyotp==2.9.0
pypng==0.20220715.0
python-decouple==3.8
python-jose==3.4.0
qrcode==7.4.2
rsa==4.9
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
celery==5.3.0
redis==4.6.0
django-celery-beat==2.5.0
